add function to get the status of a pipeline
